Refactor task management by consolidating use cases into TaskListSummaryHelper and removing unused classes
