--- lib/domain/progress_counter_use_case.dart
@@ -1,25 +1,0 @@
-import 'package:injectable/injectable.dart';
-import 'package:todoapp/data/model/task.dart';
-
-abstract class ProgressCounterUseCase {
-  double calculateProgress({required List<Task> tasks});
-}
-
-@Injectable(as: ProgressCounterUseCase)
-class ProgressCounterUseCaseImpl extends ProgressCounterUseCase {
-  @override
-  double calculateProgress({required List<Task> tasks}) {
-    int completedTasks = 0;
-    for (var task in tasks) {
-      if (task.isCompleted) {
-        completedTasks++;
-      }
-    }
-
-    if (tasks.isNotEmpty) {
-      return completedTasks / tasks.length.toDouble();
-    } else {
-      return 0.0;
-    }
-  }
-}
