--- lib/domain/should_show_share_button_use_case.dart
@@ -1,15 +1,0 @@
-import 'package:injectable/injectable.dart';
-import 'package:todoapp/data/model/task.dart';
-
-abstract class ShouldShowShareButtonUseCase {
-
-  bool shouldShowShareButton(List<Task> tasks);
-}
-
-@Injectable(as: ShouldShowShareButtonUseCase)
-class ShouldShowShareButtonUseCaseImpl extends ShouldShowShareButtonUseCase {
-  @override
-  bool shouldShowShareButton(List<Task> tasks) {
-    return tasks.any((task) => task.isCompleted == false);
-  }
-}
