+++ lib/ui/components/widgets/task/taskslist/tasks_viewmodel.dart
@@ -3,11 +3,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:todoapp/data/model/task.dart';
 import 'package:todoapp/data/model/tasks_complete_status.dart';
 import 'package:todoapp/data/todo_repository.dart';
-import 'package:todoapp/domain/format_task_list_message_use_case.dart';
-import 'package:todoapp/domain/progress_counter_use_case.dart';
-import 'package:todoapp/domain/should_show_share_button_use_case.dart';
+import 'package:todoapp/domain/task_list_summary_helper.dart';
 import 'package:todoapp/domain/tasks_comparator_use_case.dart';
-import 'package:todoapp/domain/tasks_complete_status_use_case.dart';
 import 'package:todoapp/domain/tasks_sorter_use_case.dart';
 import 'package:todoapp/ui/components/widgets/task/taskslist/tasks_screen_state.dart';
 import 'package:todoapp/util/share_message_handler.dart';
@@ -15,23 +12,17 @@ import 'package:todoapp/util/share_message_handler.dart';
 class TasksViewModel extends Cubit<TasksScreenState> {
   late TodoRepository _repository;
   late ShareMessageHandler _shareMessageHandler;
+  late TaskListSummaryHelper _taskListSummaryHelper;
 
-  ShouldShowShareButtonUseCase shouldShowShareButtonUseCase;
   TasksSorterUseCase tasksSorterUseCase;
   TasksComparatorUseCase tasksComparatorUseCase;
-  ProgressCounterUseCase progressCounterUseCase;
-  FormatTaskListMessageUseCase formatTaskListMessageUseCase;
-  TasksCompleteStatusUseCase tasksCompleteStatusUseCase;
 
   TasksViewModel({
     required TodoRepository repository,
     required ShareMessageHandler shareMessageHandler,
-    required this.shouldShowShareButtonUseCase,
+    required TaskListSummaryHelper taskListSummaryHelper,
     required this.tasksSorterUseCase,
     required this.tasksComparatorUseCase,
-    required this.progressCounterUseCase,
-    required this.formatTaskListMessageUseCase,
-    required this.tasksCompleteStatusUseCase,
   }) : super(
           const TasksScreenState(
             tasks: [],
@@ -42,6 +33,7 @@ class TasksViewModel extends Cubit<TasksScreenState> {
         ) {
     _repository = repository;
     _shareMessageHandler = shareMessageHandler;
+    _taskListSummaryHelper = taskListSummaryHelper;
   }
 
   void _onLoad() {
@@ -51,8 +43,9 @@ class TasksViewModel extends Cubit<TasksScreenState> {
   }
 
   Future<void> onCompleteButtonAction(int? checklistId) async {
-    TasksCompleteStatus status =
-        tasksCompleteStatusUseCase.checkStatus(state.tasks);
+    TasksCompleteStatus status = _taskListSummaryHelper.checkStatus(
+      tasks: state.tasks,
+    );
 
     await _repository.updateTasks(
       state.tasks,
@@ -67,22 +60,25 @@ class TasksViewModel extends Cubit<TasksScreenState> {
 
     var tasks = await _repository.getTasks(checklistId);
 
-    TasksCompleteStatus status = tasksCompleteStatusUseCase.checkStatus(tasks);
+    TasksCompleteStatus status = _taskListSummaryHelper.checkStatus(
+      tasks: tasks,
+    );
 
     emit(
       state.copyWith(
         isLoading: false,
         tasks: tasks,
         tasksCompleteStatus: status,
-        showShareIcon:
-            shouldShowShareButtonUseCase.shouldShowShareButton(tasks),
-        progress: progressCounterUseCase.calculateProgress(tasks: tasks),
+        showShareIcon: _taskListSummaryHelper.shouldShowShareButton(
+          tasks: tasks,
+        ),
+        progress: _taskListSummaryHelper.calculateProgress(tasks: tasks),
       ),
     );
   }
 
   Future<void> shareTasks({required String checklistName}) async {
-    final checklist = formatTaskListMessageUseCase.formatTaskList(
+    final checklist = _taskListSummaryHelper.formatTaskList(
       tasks: state.tasks,
     );
 
@@ -107,9 +103,10 @@ class TasksViewModel extends Cubit<TasksScreenState> {
         tasks[index] = tasks[index].copyWith(isCompleted: value);
         emit(
           state.copyWith(
-            progress: progressCounterUseCase.calculateProgress(tasks: tasks),
-            showShareIcon:
-                shouldShowShareButtonUseCase.shouldShowShareButton(tasks),
+            progress: _taskListSummaryHelper.calculateProgress(tasks: tasks),
+            showShareIcon: _taskListSummaryHelper.shouldShowShareButton(
+              tasks: tasks,
+            ),
             isLoading: false,
             tasks: tasks,
           ),
@@ -128,9 +125,10 @@ class TasksViewModel extends Cubit<TasksScreenState> {
       tasks.remove(task);
       emit(
         state.copyWith(
-            progress: progressCounterUseCase.calculateProgress(tasks: tasks),
-            showShareIcon:
-                shouldShowShareButtonUseCase.shouldShowShareButton(tasks),
+            progress: _taskListSummaryHelper.calculateProgress(tasks: tasks),
+            showShareIcon: _taskListSummaryHelper.shouldShowShareButton(
+              tasks: tasks,
+            ),
             isLoading: false,
             tasks: tasks),
       );
@@ -152,8 +150,9 @@ class TasksViewModel extends Cubit<TasksScreenState> {
     emit(
       state.copyWith(
         isLoading: false,
-        showShareIcon:
-            shouldShowShareButtonUseCase.shouldShowShareButton(tasks),
+        showShareIcon: _taskListSummaryHelper.shouldShowShareButton(
+          tasks: tasks,
+        ),
         tasks: tasks,
       ),
     );
